requirements.txt
django==1.9.3
celery==3.1.17
psycopg2==2.6.1
pytz==2016.4
django-pagination-py3

flower==0.9.0
django-debug-toolbar==1.5
django-nose==1.4.3
factory-boy==2.6.0
nose==1.3.7
gunicorn==19.4.5
ipdb==0.8.1
flake8==3.0.3